Refactor StringSet into tools package
